pyproject.toml
[project]
name = "pypub"
version = "0.0.1"
description = "A desktop Micropub client for Windows"
authors = [
    {name = "Pablo Murad", email = "pmurad@disroot.org"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyside6>=6.6.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "markdown-it-py>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "keyring>=24.0.0",
    "nh3>=0.2.14", # For HTML sanitization (Phase 2)
    "platformdirs>=4.2.0", # For Windows app paths (Phase 3)
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-qt>=4.4.0",
    "pyinstaller>=6.5.0",
    "pyinstaller-hooks-contrib>=2024.8",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
pypub = "pypub.main:main"
